more complex script test wip
